cleaner_app/requirements.txt
PyQt6
psutil
openai
python-dotenv
markdown
pygments
